+++ src/index.test.ts
@@ -115,6 +115,40 @@ describe('batchQueue', () => {
     await queue.flush()
     expect(calls).toEqual([['Joe', 'Frank', 'Bob']])
   })
+  test('should flush queue if batchBytes is reached', async () => {
+    const calls: any[] = []
+    const fn = async (records: any[]) => {
+      calls.push(records)
+    }
+    const batchSize = 5
+    const batchBytes = 8
+
+    const queue = batchQueue(fn, { batchSize, batchBytes })
+    const records = ['Joe', 'Frank', 'Bob']
+
+    for (const record of records) {
+      await queue.enqueue(record)
+    }
+    await queue.flush()
+    expect(calls).toEqual([['Joe', 'Frank'], ['Bob']])
+  })
+  test('should flush queue if batchSize is reached before batchBytes', async () => {
+    const calls: any[] = []
+    const fn = async (records: any[]) => {
+      calls.push(records)
+    }
+    const batchSize = 2
+    const batchBytes = 100
+
+    const queue = batchQueue(fn, { batchSize, batchBytes })
+    const records = ['Joe', 'Frank', 'Bob']
+
+    for (const record of records) {
+      await queue.enqueue(record)
+    }
+    await queue.flush()
+    expect(calls).toEqual([['Joe', 'Frank'], ['Bob']])
+  })
 })
 
 describe('pausable', () => {